Separate the invoke implementation
Further reduce interdependency by separating invoke implementation from
thread implementation.
