+++ mingw.mutex.h
@@ -50,9 +50,10 @@
 #endif
 #include <synchapi.h> //  For InitializeCriticalSection, etc.
 #include <errhandlingapi.h> //  For GetLastError
+#include <handleapi.h>
 
 //  Need for the implementation of invoke
-#include "mingw.thread.h"
+#include "mingw.invoke.h"
 
 #if !defined(_WIN32_WINNT) || (_WIN32_WINNT < 0x0501)
 #error To use the MinGW-std-threads library, you will need to define the macro _WIN32_WINNT to be 0x0501 (Windows XP) or higher.
